# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MemberRole = typeof MemberRole[keyof typeof MemberRole];


export const MemberRole = {
  super_admin: 'super_admin',
  admin: 'admin',
  member: 'member',
  pending: 'pending',
} as const;

export type MemberStatus = typeof MemberStatus[keyof typeof MemberStatus];


export const MemberStatus = {
  pending: 'pending',
  active: 'active',
  rejected: 'rejected',
} as const;

export interface Member {
  id: number;
  name: string;
  email: string;
  role: MemberRole;
  status: MemberStatus;
  createdAt: string;
}

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  email: string;
}

export interface InviteInput {
  /** @minLength 1 */
  code: string;
}

export interface InviteValidation {
  valid: boolean;
  /** @nullable */
  label?: string | null;
}

export type PipelineStatus = typeof PipelineStatus[keyof typeof PipelineStatus];


export const PipelineStatus = {
  new: 'new',
  interviewed: 'interviewed',
  voting: 'voting',
  bid_extended: 'bid_extended',
  not_extended: 'not_extended',
  accepted: 'accepted',
  declined: 'declined',
} as const;

export interface PipelineCount {
  status: PipelineStatus;
  count: number;
}

export type VotingRoundStatus = typeof VotingRoundStatus[keyof typeof VotingRoundStatus];


export const VotingRoundStatus = {
  open: 'open',
  closed: 'closed',
} as const;

export interface VotingRound {
  id: number;
  name: string;
  status: VotingRoundStatus;
  pnmIds: number[];
  /** @nullable */
  deadline?: string | null;
  openedAt: string;
  /** @nullable */
  closedAt?: string | null;
  voteCount: number;
}

export interface Dashboard {
  totalPnms: number;
  activePnms: number;
  pendingApprovals: number;
  pipeline: PipelineCount[];
  openRounds: VotingRound[];
  outstandingVotes: number;
}

export interface Pnm {
  id: number;
  firstName: string;
  lastName: string;
  /** @nullable */
  pronouns: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  year: string | null;
  /** @nullable */
  major: string | null;
  /** @nullable */
  minor: string | null;
  /**
     * @minimum 0
     * @maximum 4
     * @nullable
     */
  gpa: number | null;
  /** @nullable */
  photoPath?: string | null;
  status: PipelineStatus;
  archived: boolean;
  /** @nullable */
  semester?: string | null;
  /** @nullable */
  averageVote: number | null;
  voteCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface PNMInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @nullable */
  pronouns?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  major?: string | null;
  /** @nullable */
  minor?: string | null;
  /**
     * @minimum 0
     * @maximum 4
     * @nullable
     */
  gpa?: number | null;
  /** @nullable */
  photoPath?: string | null;
  status?: PipelineStatus;
  /** @nullable */
  semester?: string | null;
}

export type PNMUpdate = PNMInput & {
  firstName?: string;
  lastName?: string;
};

export interface CSVImportInput {
  /** @minLength 1 */
  csv: string;
  /** @nullable */
  semester?: string | null;
}

export interface CSVRowError {
  row: number;
  message: string;
}

export interface CSVImportResult {
  imported: number;
  errors: CSVRowError[];
}

export interface Note {
  id: number;
  pnmId: number;
  authorName: string;
  content: string;
  pinned: boolean;
  createdAt: string;
}

export interface NoteInput {
  /** @minLength 1 */
  content: string;
}

export interface NotePinUpdate {
  pinned: boolean;
}

export interface VotingRoundInput {
  /** @minLength 1 */
  name: string;
  /** @minItems 1 */
  pnmIds: number[];
  /** @nullable */
  deadline?: string | null;
}

export interface AdminVote {
  /**
     * @minimum 1
     * @maximum 5
     */
  score: number;
  memberName: string;
  createdAt: string;
}

export interface VoteResult {
  pnmId: number;
  pnmName: string;
  /** @nullable */
  average: number | null;
  voteCount: number;
  votes?: AdminVote[];
}

export type VotingRoundDetail = VotingRound & {
  results?: VoteResult[];
};

export interface VoteInput {
  pnmId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  score: number;
}

export interface VoteReceipt {
  roundId: number;
  pnmId: number;
  score: number;
  saved: boolean;
}

export interface PendingApproval {
  id: number;
  name: string;
  email: string;
  requestedAt: string;
  /** @nullable */
  inviteLabel?: string | null;
}

export type RoleUpdateRole = typeof RoleUpdateRole[keyof typeof RoleUpdateRole];


export const RoleUpdateRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface RoleUpdate {
  role: RoleUpdateRole;
}

export interface ActivityLog {
  id: number;
  actorName: string;
  action: string;
  target: string;
  createdAt: string;
}

export interface UploadInput {
  name: string;
  /** @minimum 1 */
  size: number;
  contentType: string;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}

export type SearchParameter = string;

export type YearParameter = string;

export type MajorParameter = string;

export type PipelineStatusParameter = PipelineStatus;

export type ArchivedParameter = boolean;

export type SortParameter = typeof SortParameter[keyof typeof SortParameter];


export const SortParameter = {
  name: 'name',
  year: 'year',
  major: 'major',
  status: 'status',
  score: 'score',
} as const;

export type ListPnmsParams = {
search?: SearchParameter;
year?: YearParameter;
major?: MajorParameter;
status?: PipelineStatusParameter;
archived?: ArchivedParameter;
sort?: SortParameter;
};

export type ListActivityParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};